Add files via upload
This is the R script I used for the otolith analysis. Hopefully it's understandable, I rarely write code that others have to read so it likely reads like a total hack. It's adapted from a prior project but I think everything has been switched over and/or changed to reflect this project.